Add get leaderboard functionality

+++ Core/Managers/Contracts/Models/SeasonLeaderboardPlayerData.cs
@@ -7,7 +7,7 @@ public class SeasonLeaderboardPlayerData
     public required double Assists { get; init; }
     public required double Deaths { get; init; }
     public required double KD { get; init; }
-    public required double HSP { get; init; }
+    public required decimal HSP { get; init; }
     public required double WinPercentage { get; init; }
     public required double Rating { get; init; }
 }
